@@ -485,13 +485,13 @@ class OspSimulationCase:
         fmu_file_in_case_folder: Path = (self.case_folder / fmu_file.name).resolve().absolute()
         if not fmu_file_in_case_folder.exists():
             logger.info(f"Copy FMU {fmu_file} --> {fmu_file_in_case_folder}")
-            copy2(fmu_file, self.case_folder)
+            _ = copy2(fmu_file, self.case_folder)
             # Check whether also an <fmu_name>_OspModelDescription.xml file exists.
             # If so, copy also that one.
             osp_model_description_file = fmu_file.with_name(f"{fmu_file.stem}_OspModelDescription.xml")
             if osp_model_description_file.exists():
                 logger.info(f"Copy OspModelDescription {osp_model_description_file} --> {fmu_file_in_case_folder}")
-                copy2(osp_model_description_file, self.case_folder)
+                _ = copy2(osp_model_description_file, self.case_folder)
         return fmu_file_in_case_folder
 
     def _check_components_step_size(self) -> None:
